feat: update Point class to use double for X and Y coordinates.

+++ Assignment2/Assignment2/Point.cs
@@ -2,10 +2,10 @@ namespace Assignment2;
 
 public class Point
 {
-    public int X { get; set; }
-    public int Y { get; set; }
+    public double X { get; set; }
+    public double Y { get; set; }
 
-    public Point(int x, int y)
+    public Point(double x,double y)
     {
         X = x;
         Y = y;
